% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{get_hyperparameter_samples}
\alias{get_hyperparameter_samples}
\title{Extract posterior hyperparameter samples}
\usage{
get_hyperparameter_samples(
  ibmfit,
  n_samples = 1000,
  format = c("long", "wide")
)
}
\arguments{
\item{ibmfit}{An \code{ibmfit} object.}

\item{n_samples}{Maximum number of draws.}

\item{format}{Either \code{"long"} or \code{"wide"}.}
}
\value{
A data frame.
}
\description{
Extract posterior hyperparameter samples
}
