% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{get_hyperparameter_summary}
\alias{get_hyperparameter_summary}
\title{Summarize posterior hyperparameters}
\usage{
get_hyperparameter_summary(
  ibmfit,
  n_samples = 1000,
  probs = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{ibmfit}{An \code{ibmfit} object.}

\item{n_samples}{Maximum number of draws.}

\item{probs}{Posterior probabilities.}
}
\value{
A data frame.
}
\description{
Summarize posterior hyperparameters
}
