% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_clock.R
\name{plot_diffusion}
\alias{plot_diffusion}
\title{Plot interval-specific diffusion rates}
\usage{
plot_diffusion(ibmfit, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{ibmfit}{A locally adaptive \code{ibmfit}.}

\item{probs}{Lower, center, and upper posterior probabilities.}
}
\value{
A ggplot object.
}
\description{
Plot interval-specific diffusion rates
}
