% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{print.summary_ibmfit}
\alias{print.summary_ibmfit}
\title{Print an IBM fit summary}
\usage{
\method{print}{summary_ibmfit}(x, ...)
}
\arguments{
\item{x}{A \code{summary_ibmfit} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an IBM fit summary
}
