% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{summary.ibmfit}
\alias{summary.ibmfit}
\title{Summarize an IBM fit}
\usage{
\method{summary}{ibmfit}(object, ..., n_samples = 1000, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{object}{An \code{ibmfit} object.}

\item{...}{Unused.}

\item{n_samples}{Maximum number of draws.}

\item{probs}{Posterior probabilities.}
}
\value{
An object of class \code{summary_ibmfit}.
}
\description{
Summarize an IBM fit
}
