% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{get_curve_samples}
\alias{get_curve_samples}
\title{Extract posterior curve samples}
\usage{
get_curve_samples(
  ibmfit,
  param = c("f", "fprime", "both"),
  n_samples = 1000,
  format = c("matrix", "long")
)
}
\arguments{
\item{ibmfit}{An \code{ibmfit} object.}

\item{param}{One of \code{"f"}, \code{"fprime"}, or \code{"both"}.}

\item{n_samples}{Maximum number of draws.}

\item{format}{Either \code{"matrix"} or \code{"long"}.}
}
\value{
A matrix, list of matrices, or long data frame.
}
\description{
Extract posterior curve samples
}
