% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{get_curve_summary}
\alias{get_curve_summary}
\title{Summarize posterior curves}
\usage{
get_curve_summary(ibmfit, n_samples = 1000, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{ibmfit}{An \code{ibmfit} object.}

\item{n_samples}{Maximum number of draws.}

\item{probs}{Posterior probabilities.}
}
\value{
A data frame with pointwise summaries.
}
\description{
Summarize posterior curves
}
