% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{get_curve_summary}
\alias{get_curve_summary}
\title{Summarize posterior curves}
\usage{
get_curve_summary(
  ibmfit,
  n_samples = 1000,
  probs = c(0.025, 0.1, 0.5, 0.9, 0.975)
)
}
\arguments{
\item{ibmfit}{An object of class \code{ibmfit}.}

\item{n_samples}{Number of samples to use for INLA fits, or maximum number
of MCMC draws to keep for Stan fits.}

\item{probs}{Numeric vector of posterior probabilities.}
}
\value{
A data frame with one row per parameter and time point.
}
\description{
Compute pointwise posterior summaries for the latent function and derivative.
}
