% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{get_samples}
\alias{get_samples}
\title{Extract posterior samples from an IBM fit}
\usage{
get_samples(ibmfit, param = c("f", "fprime"), n_samples = NULL, ...)
}
\arguments{
\item{ibmfit}{An object of class \code{ibmfit}.}

\item{param}{Character, either \code{"f"} or \code{"fprime"}.}

\item{n_samples}{Optional number of posterior draws to return.}

\item{...}{Additional arguments passed to \code{get_samples_inla()} for INLA fits.}
}
\value{
A matrix of posterior samples.
}
\description{
Extract posterior samples from an IBM fit
}
