% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{print.summary_ibmfit}
\alias{print.summary_ibmfit}
\title{Print an ibmfit summary}
\usage{
\method{print}{summary_ibmfit}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{summary.ibmfit()}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The input object, invisibly.
}
\description{
Print an ibmfit summary
}
